File-splitting now working with filenaming
